fix(renderer): eliminate IBL blotches from high-dynamic-range HDRI suns
Three stacked defects, all exposed by a ~97k-radiance sun (30,000:1 peak/mean):
- IBL cubemaps RGBA16F -> RGBA32F; sun texels overflowed half-float max
  (65504) into Inf/NaN that mipmapping smeared. Plus 1e6 safety clamp in
  the equirect->cube capture.
- Enable GL_TEXTURE_CUBE_MAP_SEAMLESS; face-edge clamp filtering rendered
  as face-shaped patches on the tiny high-roughness prefilter mips.
- MIP_BIAS +2 on the prefilter's filtered-importance-sampling mip: pre-
  spreads sun energy before sampling, cutting measured p95 texel error
  111% -> 3% with no energy loss (more samples provably don't help).

+++ Timefall/src/Platform/OpenGL/OpenGLTextureCube.cpp
@@ -14,7 +14,8 @@ namespace Timefall
 		: m_Size(size), m_MipLevels(mipLevels == 0 ? 1 : mipLevels)
 	{
 		glCreateTextures(GL_TEXTURE_CUBE_MAP, 1, &m_RendererID);
-		glTextureStorage2D(m_RendererID, (GLsizei)m_MipLevels, GL_RGBA16F, (GLsizei)size, (GLsizei)size);
+		// RGBA32F: HDRI suns exceed half-float max (65504) and turn Inf/NaN in 16F.
+		glTextureStorage2D(m_RendererID, (GLsizei)m_MipLevels, GL_RGBA32F, (GLsizei)size, (GLsizei)size);
 
 		glTextureParameteri(m_RendererID, GL_TEXTURE_MIN_FILTER,
 			m_MipLevels > 1 ? GL_LINEAR_MIPMAP_LINEAR : GL_LINEAR);
